Distrubuted code into packages for cleaner code
